requirements.txt
pynecone-io>=0.1.6
plotly-express
plotly
pandas